refactor: Split build environment module and add build env ABC
This will make it easier to extend which is important since the
transition to the rewritten build environment logic is likely to take
a while.
